[Fix] 인벤토리 drop 레이아웃
