+++ Assets/SHG/Scripts/UI/CraftWindow.cs
@@ -35,8 +35,11 @@ namespace SHG
       closeButton.AddToClassList("window-close-button");
       closeButton.RegisterCallback<ClickEvent>(this.OnClickCloseButton);
       this.Add(closeButton);
+      var scrollView = new ScrollView(ScrollViewMode.Vertical);
+      scrollView.AddToClassList("item-container-scroll-view");
       this.productListWindow = new ProductListWindow();   
-      this.Add(this.productListWindow);
+      scrollView.Add(this.productListWindow);
+      this.Add(scrollView);
       this.recipeWindow = new RecipesWindow();
       this.Add(this.recipeWindow);
     }
