+++ Assets/SHG/Scripts/UI/InventoryStyle.uss
@@ -4,7 +4,6 @@
   height: 70%;
   left: 60%;
   top: 10%;
-  padding-top: 100px;
   position: absolute;
 }
 
@@ -19,22 +18,12 @@
 
 .inventory-window-tab-button
 {
-  border-color: green;
-  border-width: 1px;
   color: black;
   font-size: 20px;
   margin: 5px;
 }
 
-.item-container-scroll-view
-{
-  width: 100%;
-  height: 100%;
-  align-items: center;
-  justify-content: flex-start;
-}
-
-#item-storage-window-container
+#item-locker-window-container
 {
   width: 50%;
   height: 70%;
@@ -44,7 +33,7 @@
   position: absolute;
 }
 
-#item-storage-item-container
+#item-locker-item-container
 {
   width: 100%;
   height: 80%;
