+++ Assets/SHG/Scripts/UI/ItemStorageStyle.uss.meta
