+++ Assets/SHG/Scripts/UI/ItemStorageWindow.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7d91b64654fef4c23a3242ad1bad6849
+guid: 025806470cffda0bd82990c21d1c389a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
