+++ Assets/SHG/Scripts/UI/MainUIPlaceHolder.uss
@@ -24,7 +24,7 @@
 .window-tab-container
 {
   width: 90%;
-  height: 80%;
+  height: 100%;
   border-color: green;
   border-width: 2px;
 }
