+++ Assets/SHG/Scripts/UI/QuickSlotStyle.uss
@@ -1,12 +1,14 @@
 #quick-slot-window-container
 {
   position: absolute;
-  left: 40%;
+  left: 30%;
   bottom: 5%;
   flex-direction: row;
   align-items: center;
+  justify-content: center;
   border-color: black;
   border-width: 2px;
+  max-width: 40%;
 }
 
 .quick-slot-window-item-box
